requirements.txt
numpy
opencv-python
scikit-learn
Pillow
